refactor some things to not use context
