+++ test/integration/db/integration-db.js
@@ -1,0 +1,3 @@
+const { env } = process;
+
+module.exports = env.DANGER_INTEGRATION_DATABASE;
